flaskApp/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
pika==1.3.2
jsonschema==4.17.3
python-dotenv==1.0.0
pyyaml==6.0
kubernetes==26.1.0